+++ Game/BatStates.cpp
@@ -131,7 +131,7 @@ void BatDieState::Tick()
 	}
 	
 	//Dissolve
-	float frameTime = animInstance->GetTotalFrame();
+	float frameTime = static_cast<float>(animInstance->GetTotalFrame());
 	frameTime /= 30;
 	m_fDissolveTimer += TIMER->GetDeltaTime();
 	float t = m_fDissolveTimer / frameTime;
