+++ Game/MageMovement.cpp
@@ -139,7 +139,7 @@ void MageMovement::ChangeState(shared_ptr<StateBase> _state)
 
 	if (currentState && !currentState->IsInterruptible() && currentState->IsPlaying())
 	{
-		if (!_state->GetId() == ENEMY_S_DEATH)
+		if (!(_state->GetId() == static_cast<UINT>(ENEMY_S_DEATH)))
 		{
 			return;
 		}
